services/data_pipeline/requirements.txt
# requirements.txt (Basis ohne TensorFlow)
# Core
nautilus_trader>=1.200.0
numpy>=1.26.0,<2.0.0
pandas>=2.1.0
scipy>=1.11.0

# Data & Storage
psycopg2-binary>=2.9.9
redis>=5.0.0
pyarrow>=14.0.0
fastparquet>=2023.10.0

# ML (ohne TensorFlow/PyTorch)
scikit-learn>=1.3.0
statsmodels>=0.14.0

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
tqdm>=4.66.0